refactor(card): reajustando as pastas de card de projetos
