+++ src/assets/icons/MediumIcon.js
