+++ src/components/Cards/CardImage/styles/index.js
